Drinker create

+++ src/main/java/pl/markowski/konrad/drinkingapp/web/model/ContainerModel.java
@@ -3,9 +3,9 @@ package pl.markowski.konrad.drinkingapp.web.model;
 //public record ContainerModel(String name, int volume) {
 public class ContainerModel {
     protected String name;
-    protected int volume;
+    protected double volume;
 
-    public ContainerModel(String name, int volume) {
+    public ContainerModel(String name, double volume) {
         this.name = name;
         this.volume = volume;
     }
